+++ tests/numbers.scm
@@ -434,6 +434,9 @@
         (t.is (string->number "1e2" 16) 482)
         (t.is (string->number "1e2") 100)
 
+        ;; #326
+        (t.is (string->number "2" 1) +nan.0)
+
         ;; accept radix
         (t.is (string->number "A" 16) 10)
         (t.is (string->number "10" 8) 8)